Drop findIndirect() algorithm
According to cppreference.com, std::invoke is supposed to deal with one
level of indirection automatically (e.g. if the first parameter is a
smart pointer), so we should not need findIndirect() for that. What
findIndirect() effectively does is enabling the second level of
indirection, and that kind of stuff better be spelled out in the code
instead of being magically consumed by an algorithm.