Quotient::rangeTo()
This is a limited substitute for std::ranges::to() that is still not
supported by Xcode. With plenty of other range adaptors not in Apple
libc++ either, this can only be used sparingly but it's better this way
than leaving a TODO in every place std::ranges::to() could have been
used.