Emit finished() and delete remotely cancelled verification session
KeyVerificationSession emits finished() and self-deletes on any wrong
situation but only changes state if a remote device simply cancels the
verification, as if it were not an endgame. This commit pulls the common
code from cancelVerification() into handleCancel() and calls that new
function when KeyVerificationCancelEvent is received. It also makes sure
now not to call cancelVerification() twice with different error codes.