KVS::handleReady: Move setState(READY) to the end
Other handlers set the state after doing their stuff, not before; and
this is actually more correct. Also, this particular READY state is
currently very brief because the library will most likely immediately
initiate the SAS method because it's the only one it supports, there's
no need to ask the user about it.