Replace JobHandle public constructor with a factory method
The newly introduced JobHandle<>::createFrom() creates a new job object
and immediately wraps it into a JobHandle. The code that uses JobHandle
can no more create them from bare pointers (the previous commit
explained why it's problematic). Instead, you either create a job object
explicitly, call non-templated Connection::run() on it and live with job
pointers; or you use Connection::callApi() or JobHandle::createFrom(),
get a job handle and use it either via a pointer interface or as a
future. In case of JH::createFrom(), you should also call templated
Connection::run() to actually start the job, where/when appropriate.

This change also forced to change the return type of
Connection::getThumbnail(), so that Avatar code continued building.
Previously this code was re-creating a job handle from a raw pointer
returned by getThumbnail(), causing the job handle previously created by
callApi to become dysfunctional. That first job handle was deleted at
the end of getThumbnail() but Qt still complained about repeated
continuation attachment because the underlying promise is the same. Even
though that warning was benign in this case, I've had other cases of
repeated attachments in the past that weren't very easy to track down
since by the moment the second attachment occurs there's no knowing
about the whereabouts of the first one.

Another case of turning raw pointers to JobHandles was in tests for
QOlmAccount; for that reason, QOlmAccount::createUploadKeyRequest() now
also returns JobHandle, in the unchanged code in tests now picks the
templated overload of Connection::run().