Deprecate Expected; migrate to std::expected
Most of the changes are due to std::expected constructor being explicit
if the value is of the error type, rather than the value type - places
where an error is returned have to use std::unexpected now across the
board. There's also no move_value_or, but that's an even smaller issue.