.clang-format: align closer with Qt formatting
Yes, another rehash of ClangFormat, should've been done on the first
time, sorry. Changes: braces on the new line for classes/structs (back
again, after using them on the same line for some time - both Qt and KDE
put them on the new line); pointer/reference alignment is now on the
right, not left (brand new change, it was on the left since very early
in the lib development but again, both Qt and KDE align it on the
right).

[skip ci]