EventMetaType: tighter checks on type clashes
Previous logic relied on comparison of string literal addresses, and
even that logic left warnings in the logs about RoomJoinRulesEvent that
didn't get it's QUOTIENT_API attribute when you build quotest or a
client and link libQuotient as a shared library. Instead of a mere
warning, that case should've been caught by QUO_ALARM_X a few lines
earlier because it was literally the same class getting registered twice
from two different translation units, the immediate cause of #413; but
the string literals with the name of this class ended up placed
separately in both translation units too, so QUO_ALARM_X wasn't
triggered.

The refactored code makes use of std::type_info and (static version of)
typeid(), i.e. the now-standard tools to compare and order types. Not
only this is more robust against same class registration, it also
provides the fully-qualified name (with namespaces etc.) so
JoinRulesEvent and Quotient::JoinRulesEvent would now be distinct
classes (meaning - it will still cause a warning, not an assertion
failure, to register both).

Because std::type_info::name() returns a mangled type name when compiled
with GCC and Clang, I got to use a piece of private Qt API that extracts
the right name by certain manipulations involving Q_FUNC_INFO. This
trick is, in fact, rather old and well-known
(see https://stackoverflow.com/a/66551751/16842332, e.g.) and Qt uses
it, increasingly sophisticated, since before Qt 6 (see https://
github.com/qt/qtbase/commit/
33cd680ddbaccf6139e215d851a39e657ae36394#diff-6c4e4eae998bb9654f772cc4939fa351aaf2ccebaa05043968b76f7ac9bc24a5R2288
for the still succinct precursor to the current code) - so there's hope
this internal API won't significantly change. It would be ideal to use
QMetaType::fromType<>() instead but it doesn't work quite well for
incompletely defined types with out-of-line constructors (so, anything
derived from RoomEvent).

This commit (intentionally) breaks the build, to expose the current
issue in the code. The next commit will fix it again.