Use smaller number when disabling timeouts for sync jobs
This number is too big when we use it inside QObject::startTimer- for
various reasons, such as when it internally converts it to nanoseconds
and overflows it's integer type. This causes some consumers to emit the
following log message:

QObject::startTimer: Timers cannot have negative intervals

To avoid this, we can choose a smaller value for the job timeout and
still have the same result of effectively disabling the timeout.

Fixes #885