ConnectionEncryptionData: Only delete once in saveDevicesList()
This removes part of what takes this function so long to execute (#888)
because we end up calling delete twice, and one is more expensive than
the other. The first one culls all tracked_device rows that match on
Matrix ID, and then we do it *again* but to be more specific based on
each of that user's devices.

Depending on how many devices we have tracked (which might be a lot,
based on some of the database sizes we've seen) this can take so long it
freezes the application UI. We can remove the second, more specific
query and reduce - but not remove - the freezing.