Drop lowerBoundMemberIndex()
It's too small a piece of functionality, with the API accidentally
hardwired to Quotient::MemberSorter on top of it. Ended up barely used,
and with std::flat_map on the horizon (maybe in a couple of years still,
but at least standardised), having extra wrappers around std::
(ranges::)lower_bound() doesn't make much sense in the long run.