Fix qt warning about too high timer intervals
Qt warns that it can't set timers for 1000 years and that they will be clamped to "about 24 days".
For simplicity, set the interval to three weeks, which is close enough to the maximum and simple enough to write;
the exact value doesn't matter to us anyway.

(cherry picked from commit 434d10a1ec1a594aefcc99b00d7db1161bf1dfbc)