CI: Use GCC 14 and Clang 19
The new versions are necessary to enable std::expected. In fact, Clang pipeline
has already been using stdlibc++ 14 because it's the latest standard library
available in the Ubuntu image now; but with Clang 18 the library #ifdef's out
the definition of std::expected because the compiler doesn't advertise
sufficient concepts support (201709; the library wants 202202).