JobHandle: make sure BoundFn never holds a reference
In most cases continuation is passed as an rvalue so BoundFn moves and
stores it as it should. However, if a continuation is an lvalue - e.g.,
a closure object created on the stack and then passed to then() or
onResult() - then BoundFn uses "auto&" as its template parameter and
once the closure object goes out of the scope it was created in, the
continuation is left with a dangling reference, as #910 shows.

(cherry picked from commit 689584cd9bbe50f1ab28fff403c4127f9744d286)