Don't store converted content in state events
Instead, both current and previous content is now produced on the fly,
on each content() and prevContent() call. This is obviously suboptimal
in case of repeated content() calls but saves both time and space when
most data from state events are not actually requested - which seems to
be a common case, as clients rarely visualise each and every part of the
state. When benchmarking on Quotest, and assuming the next commit (that
optimises RoomMemberEvent, the most frequent state event of all)
applied, the slowdown was about 2.5% and even that was within
the deviation range.