Convert leaveRoom() and forgetRoom() to futures
In particular, Connection::forgetRoom() has long been a strong case to
use futures as it performs two operations in sequence. The ony non-
trivial overload of JobHandle::rewrap() that was meant for that, still
could not be used though; the reason being, leaveRoom() already returns
QFuture rather than JobHandle so that it could return early. Probably
that rewrap() thing was ultimately a bit of overengineering...